fix: seperate room state components into individual files
