+++ meet/app/r/[name]/page.tsx
@@ -1,31 +1,12 @@
 'use client';
-import {
-  LiveKitRoom,
-  LocalUserChoices,
-  VideoConference,
-  formatChatMessageLinks,
-  useToken,
-} from '@livekit/components-react';
-
-import {
-  ExternalE2EEKeyProvider,
-  LogLevel,
-  Room,
-  RoomConnectOptions,
-  RoomOptions,
-  VideoCodec,
-  VideoPresets,
-} from 'livekit-client';
-
 import '@livekit/components-styles';
 import '@livekit/components-styles/prefabs';
 
+import { decodePassphrase, encodePassphrase, randomString } from '~/utils/lksdk';
+import { LocalUserChoices } from '@livekit/components-react';
+import { ActiveRoom } from '~/components/ActiveRoom';
 import dynamic from 'next/dynamic';
-import { useMemo, useState } from 'react';
-import { DebugMode } from '~/components/Debug';
-import { decodePassphrase, encodePassphrase, randomString, useServerUrl } from '~/utils/lksdk';
-import router from 'next/router';
-import { useRouter, useSearchParams } from 'next/navigation';
+import { useState } from 'react';
 
 const PreJoinNoSSR = dynamic(
   async () => {
@@ -34,7 +15,7 @@ const PreJoinNoSSR = dynamic(
   { ssr: false },
 );
 
-export default function Page({ params }: { params: { name: string } }) {
+export default function LivekitRoom({ params }: { params: { name: string } }) {
   const roomName = params.name;
 
   const e2eePassphrase =
@@ -48,17 +29,11 @@ export default function Page({ params }: { params: { name: string } }) {
     }
     setPreJoinChoices(values);
   }
+
   return (
     <main data-lk-theme="default" className="h-screen">
       {roomName && !Array.isArray(roomName) && preJoinChoices ? (
-        <ActiveRoom
-          roomName={roomName}
-          userChoices={preJoinChoices}
-          onLeave={() => {
-            console.log('leaving room');
-            router.push('/r');
-          }}
-        />
+        <ActiveRoom roomName={roomName} userChoices={preJoinChoices} />
       ) : (
         <div className="flex flex-col justify-center items-center h-full">
           <h1 className="text-4xl font-semibold text-foreground font-display mb-4">Kitt2</h1>
@@ -79,93 +54,3 @@ export default function Page({ params }: { params: { name: string } }) {
     </main>
   );
 }
-
-type ActiveRoomProps = {
-  userChoices: LocalUserChoices;
-  roomName: string;
-  region?: string;
-  onLeave?: () => void;
-};
-const ActiveRoom = ({ roomName, userChoices, onLeave }: ActiveRoomProps) => {
-  const token = useToken(process.env.NEXT_PUBLIC_LK_TOKEN_ENDPOINT, roomName, {
-    userInfo: {
-      identity: userChoices.username,
-      name: userChoices.username,
-    },
-  });
-
-  const searchParams = useSearchParams();
-  const region = searchParams.get('region');
-  const hq = searchParams.get('hq');
-  const codec = searchParams.get('codec') ?? 'vp9';
-
-  const liveKitUrl = useServerUrl(region as string | undefined);
-
-  const worker =
-    typeof window !== 'undefined' &&
-    userChoices.e2ee &&
-    new Worker(new URL('livekit-client/e2ee-worker', import.meta.url));
-
-  const e2eeEnabled = !!(userChoices.e2ee && worker);
-  const keyProvider = new ExternalE2EEKeyProvider();
-
-  const roomOptions = useMemo((): RoomOptions => {
-    return {
-      videoCaptureDefaults: {
-        deviceId: userChoices.videoDeviceId ?? undefined,
-        resolution: hq === 'true' ? VideoPresets.h2160 : VideoPresets.h720,
-      },
-      publishDefaults: {
-        dtx: false,
-        videoSimulcastLayers:
-          hq === 'true'
-            ? [VideoPresets.h1080, VideoPresets.h720]
-            : [VideoPresets.h540, VideoPresets.h216],
-        red: !e2eeEnabled,
-        videoCodec: codec as VideoCodec | undefined,
-      },
-      audioCaptureDefaults: {
-        deviceId: userChoices.audioDeviceId ?? undefined,
-      },
-      adaptiveStream: { pixelDensity: 'screen' },
-      dynacast: true,
-      e2ee: e2eeEnabled
-        ? {
-            keyProvider,
-            worker,
-          }
-        : undefined,
-    };
-  }, [userChoices, hq, codec]);
-
-  const room = useMemo(() => new Room(roomOptions), []);
-
-  if (e2eeEnabled) {
-    keyProvider.setKey(decodePassphrase(userChoices.sharedPassphrase));
-    room.setE2EEEnabled(true);
-  }
-  const connectOptions = useMemo((): RoomConnectOptions => {
-    return {
-      autoSubscribe: true,
-    };
-  }, []);
-
-  return (
-    <>
-      {liveKitUrl && (
-        <LiveKitRoom
-          room={room}
-          token={token}
-          serverUrl={liveKitUrl}
-          connectOptions={connectOptions}
-          video={false}
-          audio={userChoices.audioEnabled}
-          onDisconnected={onLeave}
-        >
-          <VideoConference chatMessageFormatter={formatChatMessageLinks} />
-          <DebugMode logLevel={LogLevel.info} />
-        </LiveKitRoom>
-      )}
-    </>
-  );
-};
