+++ meet/utils/server.ts
@@ -1,0 +1,27 @@
+import { RoomServiceClient } from 'livekit-server-sdk';
+
+export function getRoomClient(): RoomServiceClient {
+  checkKeys();
+  return new RoomServiceClient(getLiveKitURL());
+}
+
+export function getLiveKitURL(region?: string | string[]): string {
+  let targetKey = 'LIVEKIT_WS_URL';
+  if (region && !Array.isArray(region)) {
+    targetKey = `LIVEKIT_WS_URL_${region}`.toUpperCase();
+  }
+  const url = process.env[targetKey];
+  if (!url) {
+    throw new Error(`${targetKey} is not defined`);
+  }
+  return url;
+}
+
+function checkKeys() {
+  if (typeof process.env.LIVEKIT_API_KEY === 'undefined') {
+    throw new Error('LIVEKIT_API_KEY is not defined');
+  }
+  if (typeof process.env.LIVEKIT_API_SECRET === 'undefined') {
+    throw new Error('LIVEKIT_API_SECRET is not defined');
+  }
+}
